Add UART option (with a #define)
Added uart option which uses an I/O line as a uart transmit line, running at 9600 baud.  Currently selected Bypass button line for this (in WA1RCT mode).  All characters that are sent to the display are sent out the UART.  the output is RCCC:xxxxxCRLF  where RCCC is the Row, Column to display the values, the x's are the characters (up to 10), and ends with a carriage return/line feed.